refactor x11 cursor changing code into x11/cursor.rs
