+++ src/x11/mod.rs
@@ -3,3 +3,5 @@ use xcb_connection::XcbConnection;
 
 mod window;
 pub use window::*;
+
+mod cursor;
